% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prismagui.R
\name{prismagui}
\alias{prismagui}
\title{Main prismutils GUI}
\usage{
prismagui()
}
\description{
Open the GUI containing utilities for PRISMA
 (at this stage, the possibility to read, visualise and export PRISMA
 orbits).
}
\seealso{
\code{\link{read_prisma_orbit}} to import XML files from the 
command line.
}
\author{
Luigi Ranghetti, phD (2020) <ranghetti.l@irea.cnr.it>
}
