% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prisma_orbit.R
\name{read_prisma_orbit}
\alias{read_prisma_orbit}
\title{Read PRISMA orbit XML files}
\usage{
read_prisma_orbit(xmlpath, format = "data.frame")
}
\arguments{
\item{xmlpath}{Character value with the path of the input XML file.}

\item{format}{(optional) One of the following options, determining the 
output format:
- `data.frame`: a data frame containing the information in tabular format;
- `polygon`: a `sf POLYGON` spatial object with the geometry of the tile;
- `point`: a `sf POINT` spatial object with the geometry of the reference 
    positions.}
}
\value{
The input orbit information (see argument `format`).
}
\description{
Read XML files containing the PRISMA orbit propagations
 (the format of this XML files is fixed and defined by Telespazio).
}
\seealso{
\code{\link{prismagui}} to read XML files from the GUI.
}
\author{
Luigi Ranghetti, phD (2020) <ranghetti.l@irea.cnr.it>
}
